ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
asyncpg==0.30.0
python-dotenv==1.1.0
langchain==0.3.0
langchain-ollama==0.3.0
langchain-core==0.3.0
google-generativeai==0.8.0
numpy==2.2.0
pydantic==2.10.0
loguru==0.7.3
httpx==0.28.0
